Changed structure of file uploading. Created File repository

+++ src/main/java/com/telerik/carpoolingapplication/services/FileServiceImpl.java
@@ -1,7 +1,8 @@
 package com.telerik.carpoolingapplication.services;
 
 import com.telerik.carpoolingapplication.models.UserDTO;
-import com.telerik.carpoolingapplication.models.constants.Messages;
+import com.telerik.carpoolingapplication.models.constants.Constants;
+import com.telerik.carpoolingapplication.repositories.FileRepository;
 import com.telerik.carpoolingapplication.repositories.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,39 +14,55 @@ import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
+import java.util.Objects;
 import java.util.UUID;
 
 @Transactional
 @Service
 public class FileServiceImpl implements FileService {
-    public static final String STORAGE_ROUTE = "/Users/jiwkojelew/Telerik/Final Project/carpooling-application/src/main/resources/static/images/users/";
     private UserRepository userRepository;
+    private FileRepository fileRepository;
 
-    public FileServiceImpl(UserRepository userRepository) {
+    public FileServiceImpl(UserRepository userRepository, FileRepository fileRepository) {
         this.userRepository = userRepository;
+        this.fileRepository = fileRepository;
     }
 
     @Override
     public void storeFile(int userId, MultipartFile file) throws IOException {
+        if(file.getSize() > 5242880) {
+            throw new IllegalArgumentException("File should be less than 5MB");
+        }
+        if (isFileFormatInvalid(file)) {
+            throw new IllegalArgumentException("Invalid file format");
+        }
         if (userRepository.getById(userId) == null) {
-            throw new IllegalArgumentException(String.format(Messages.USER_NOT_FOUND, userId));
+            throw new IllegalArgumentException(String.format(Constants.USER_NOT_FOUND, userId));
         }
         UserDTO user = userRepository.getById(userId);
-        if (!user.getAvatarUri().isEmpty() && !user.getAvatarUri().contains("default")) {
+        if (!user.getAvatarUri().isEmpty() && !user.getAvatarUri().contains(Constants.DEFAULT_USER_IMAGE_NAME)) {
             Files.delete(Paths.get(user.getAvatarUri()));
         }
-        String fileName = getFileName(file);
-        String filePath = STORAGE_ROUTE + fileName;
+
+        String fileName = changeFileName(file);
+        String filePath = Constants.STORAGE_ROUTE + fileName;
 
         try (InputStream inputStream = file.getInputStream()) {
             Files.copy(inputStream, Paths.get(filePath), StandardCopyOption.REPLACE_EXISTING);
-            userRepository.saveImage(userId, filePath);
+            fileRepository.saveImage(userId, filePath);
         } catch (IOException e) {
             throw new IOException("Failed to store file " + fileName, e);
         }
     }
 
-    private String getFileName(MultipartFile file) {
+    private boolean isFileFormatInvalid(MultipartFile file) {
+        return !Objects.equals(file.getContentType(), "image/jpeg") &&
+                !Objects.equals(file.getContentType(), "image/jpg") &&
+                !Objects.equals(file.getContentType(), "image/png") &&
+                !Objects.equals(file.getContentType(), "image/gif");
+    }
+
+    private String changeFileName(MultipartFile file) {
         String originalFileName = StringUtils.cleanPath(file.getOriginalFilename());
         int index = originalFileName.lastIndexOf('.');
         return (UUID.randomUUID()) + originalFileName.substring(index);
